requirements.txt
torchvision~=0.8.1
setuptools~=52.0.0